refactor: move compareApiKeys function and update imports for user authentication
