+++ web-app/src/lib/auth/utils.ts
@@ -1,9 +1,9 @@
+"use server";
+
 import { headers } from "next/headers";
 import { redirect } from "next/navigation";
 
-import { getEnvSecrets } from "@/config/env.config";
-import { auth, Session } from "@/lib/auth/auth";
-import { createHash } from "@/lib/utils";
+import { auth, Session, User } from "@/lib/auth/auth";
 
 export async function requireAuthentication(): Promise<{
   session: Session;
@@ -19,7 +19,14 @@ export async function requireAuthentication(): Promise<{
   return { session };
 }
 
-export const compareApiKeys = (apiKey: string) => {
-  const envApiKey = getEnvSecrets().ADMIN_KEY;
-  return createHash(apiKey) === createHash(envApiKey);
-};
+export async function getAuthenticatedUser(): Promise<User | null> {
+  const session = await auth.api.getSession({
+    headers: await headers(),
+  });
+
+  if (!session) {
+    return null;
+  }
+
+  return session.user;
+}
